+++ src/lsp.ts
@@ -8,7 +8,7 @@ import type {
 	Position as NPosition,
 	MarkupContent,
 } from 'vscode-languageserver-types';
-import type {IWikitextModel} from './wikilint.ts';
+import type {IWikitextModel} from './linter.ts';
 
 const iPositionToNPosition = ({lineNumber, column}: IPosition): NPosition => ({
 	line: lineNumber - 1,